requirements.txt
Django==2.1.4
django-cors-headers==3.0.2
djangorestframework==3.9.4
Markdown==3.1.1
pytz==2019.1
